pyproject.toml
[project]
name = "flightai-airline-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=6.2.0",
    "openai>=2.14.0",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
]
